# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 1.0.0
  description: STACK Service Desk AI Solution API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /v1/auth/login:
    post:
      operationId: login
      tags: [auth]
      summary: Login with credentials
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LoginInput"
      responses:
        "200":
          description: Login successful
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthToken"
...
  /v1/logs/audit:
    get:
      operationId: getAuditLogs
      tags: [logs]
      summary: Paginated audit logs
      parameters:
        - name: event_type
          in: query
          schema:
            type: string
        - name: actor_type
          in: query
          schema:
            type: string
        - name: use_case
          in: query
          schema:
            type: string
        - name: date_from
          in: query
          schema:
            type: string
        - name: date_to
          in: query
          schema:
            type: string
        - name: search
          in: query
          schema:
            type: string
        - name: page
          in: query
          schema:
            type: integer
            default: 1
        - name: limit
          in: query
          schema:
            type: integer
            default: 20
      responses:
        "200":
          description: Audit logs
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuditLogListResponse"
  /v1/logs/api-calls:
    get:
      operationId: getApiCallLogs
      tags: [logs]
      summary: API call logs
      parameters:
        - name: api_name
          in: query
          schema:
            type: string
        - name: date_from
          in: query
          schema:
            type: string
        - name: date_to
          in: query
          schema:
            type: string
        - name: page
          in: query
          schema:
            type: integer
            default: 1
        - name: limit
          in: query
          schema:
            type: integer
            default: 20
      responses:
        "200":
          description: API call logs
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiCallLogListResponse"
  /v1/logs/powershell:
    get:
      operationId: getPowershellLogs
      tags: [logs]
      summary: PowerShell execution logs
      parameters:
        - name: date_from
          in: query
          schema:
            type: string
        - name: date_to
          in: query
          schema:
            type: string
        - name: page
          in: query
          schema:
            type: integer
            default: 1
        - name: limit
          in: query
          schema:
            type: integer
            default: 20
      responses:
        "200":
          description: PowerShell logs
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PsLogListResponse"
...
  /v1/admin/thresholds/{id}:
    patch:
      operationId: updateThreshold
      tags: [admin]
      summary: Update confidence threshold
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ThresholdUpdate"
      responses:
        "200":
          description: Threshold updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ConfidenceThreshold"
...
  /v1/admin/sla-configs/{id}:
    patch:
      operationId: updateSlaConfig
      tags: [admin]
      summary: Update SLA config
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SlaConfigUpdate"
      responses:
        "200":
          description: SLA config updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SlaConfig"
...
  /v1/admin/groups/{id}:
    patch:
      operationId: updateAgentGroup
      tags: [admin]
      summary: Update agent group
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AgentGroupUpdate"
      responses:
        "200":
          description: Group updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AgentGroup"
  /v1/admin/agents:
    get:
      operationId: listAgents
      tags: [admin]
      summary: List agents
      responses:
        "200":
          description: Agents list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Agent"
    post:
      operationId: createAgent
      tags: [admin]
      summary: Add agent
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AgentInput"
      responses:
        "201":
          description: Agent created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Agent"
  /v1/admin/agents/{id}:
    patch:
      operationId: updateAgent
      tags: [admin]
      summary: Update agent
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AgentUpdate"
      responses:
        "200":
          description: Agent updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Agent"
...
  /v1/chat/message:
    post:
      operationId: sendChatMessage
      tags: [chat]
      summary: Send chat message (Google Chat Bot integration point)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ChatMessageInput"
      responses:
        "200":
          description: Chat response
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ChatResponse"
components:
  schemas:
...
    LoginInput:
      type: object
      required: [email, password]
      properties:
        email:
          type: string
        password:
          type: string
    AuthToken:
      type: object
      required: [token, agent]
      properties:
        token:
          type: string
        agent:
          $ref: "#/components/schemas/Agent"
    Agent:
      type: object
      required: [agent_id, email, full_name, role, is_active]
      properties:
        agent_id:
          type: string
        email:
          type: string
        full_name:
          type: string
        role:
          type: string
          enum: [admin, agent, readonly]
        freshservice_agent_id:
          type: ["string", "null"]
        is_active:
          type: boolean
        created_at:
          type: string
    AgentInput:
      type: object
      required: [email, full_name, role]
      properties:
        email:
          type: string
        full_name:
          type: string
        role:
          type: string
          enum: [admin, agent, readonly]
        freshservice_agent_id:
          type: string
    AgentUpdate:
      type: object
      properties:
        full_name:
          type: string
        role:
          type: string
          enum: [admin, agent, readonly]
        is_active:
          type: boolean
...
    AuditLog:
      type: object
      required: [log_id, event_type, actor, actor_type, created_at]
      properties:
        log_id:
          type: string
        ticket_id:
          type: ["string", "null"]
        event_type:
          type: string
        actor:
          type: string
        actor_type:
          type: string
          enum: [user, ai, agent, system]
        details:
          type: object
        ip_address:
          type: ["string", "null"]
        created_at:
          type: string
    AuditLogListResponse:
      type: object
      required: [logs, total, page, limit]
      properties:
        logs:
          type: array
          items:
            $ref: "#/components/schemas/AuditLog"
        total:
          type: integer
        page:
          type: integer
        limit:
          type: integer
    ApiCallLog:
      type: object
      required: [api_log_id, api_name, endpoint, method, response_status, called_at]
      properties:
        api_log_id:
          type: string
        ticket_id:
          type: ["string", "null"]
        api_name:
          type: string
        endpoint:
          type: string
        method:
          type: string
        request_payload:
          type: ["object", "null"]
        response_status:
          type: integer
        response_payload:
          type: ["object", "null"]
        duration_ms:
          type: ["integer", "null"]
        called_at:
          type: string
    ApiCallLogListResponse:
      type: object
      required: [logs, total, page, limit]
      properties:
        logs:
          type: array
          items:
            $ref: "#/components/schemas/ApiCallLog"
        total:
          type: integer
        page:
          type: integer
        limit:
          type: integer
    PsLog:
      type: object
      required: [execution_id, ticket_id, script_name, execution_status, executed_at]
      properties:
        execution_id:
          type: string
        ticket_id:
          type: string
        resolution_id:
          type: ["string", "null"]
        device_name:
          type: ["string", "null"]
        device_ip:
          type: ["string", "null"]
        script_name:
          type: string
        script_content:
          type: ["string", "null"]
        execution_status:
          type: string
          enum: [success, failed, timeout, partial]
        output_log:
          type: ["string", "null"]
        executed_at:
          type: string
        duration_seconds:
          type: ["integer", "null"]
    PsLogListResponse:
      type: object
      required: [logs, total, page, limit]
      properties:
        logs:
          type: array
          items:
            $ref: "#/components/schemas/PsLog"
        total:
          type: integer
        page:
          type: integer
        limit:
          type: integer
...
    ConfidenceThreshold:
      type: object
      required: [threshold_id, use_case, auto_resolve_min, review_after_min]
      properties:
        threshold_id:
          type: string
        use_case:
          type: string
        auto_resolve_min:
          type: number
        review_after_min:
          type: number
        updated_at:
          type: ["string", "null"]
    ThresholdUpdate:
      type: object
      properties:
        auto_resolve_min:
          type: number
        review_after_min:
          type: number
    SlaConfig:
      type: object
      required: [sla_id, use_case, priority, resolution_hours, warning_threshold_percent]
      properties:
        sla_id:
          type: string
        use_case:
          type: string
        priority:
          type: string
        resolution_hours:
          type: integer
        warning_threshold_percent:
          type: number
        updated_at:
          type: ["string", "null"]
    SlaConfigUpdate:
      type: object
      properties:
        resolution_hours:
          type: integer
        warning_threshold_percent:
          type: number
    AgentGroupMember:
      type: object
      required: [member_id, agent_id, full_name, email, priority_order]
      properties:
        member_id:
          type: string
        agent_id:
          type: string
        full_name:
          type: string
        email:
          type: string
        priority_order:
          type: integer
    AgentGroup:
      type: object
      required: [group_id, group_name, use_case, assignment_mode]
      properties:
        group_id:
          type: string
        group_name:
          type: string
        use_case:
          type: string
        assignment_mode:
          type: string
          enum: [round_robin, first_available, priority]
        freshservice_group_id:
          type: ["string", "null"]
        members:
          type: array
          items:
            $ref: "#/components/schemas/AgentGroupMember"
    AgentGroupUpdate:
      type: object
      properties:
        assignment_mode:
          type: string
          enum: [round_robin, first_available, priority]
...
    ChatMessageInput:
      type: object
      required: [user_id, session_id, message, platform]
      properties:
        user_id:
          type: string
        session_id:
          type: string
        message:
          type: string
        platform:
          type: string
          enum: [google_chat, web_dashboard]
        context:
          type: object
    ChatResponse:
      type: object
      required: [reply, requires_input, session_id]
      properties:
        reply:
          type: string
        ticket_id:
          type: ["string", "null"]
        action_taken:
          type: ["string", "null"]
        requires_input:
          type: boolean
        input_prompt:
          type: ["string", "null"]
        session_id:
          type: string