# lib/api-zod/src/generated/types/agentUpdate.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * STACK Service Desk AI Solution API
 * OpenAPI spec version: 1.0.0
 */
import type { AgentUpdateRole } from './agentUpdateRole';

export interface AgentUpdate {
  full_name?: string;
  role?: AgentUpdateRole;
  is_active?: boolean;
}